refactor(agent-core): move HookEngine to sessions
