+++ packages/agent-core/src/config/schema.ts
@@ -1,4 +1,4 @@
-import { HOOK_EVENT_TYPES } from '#/agent/hooks/types';
+import { HOOK_EVENT_TYPES } from '../session/hooks/types';
 import { parsePattern } from '#/agent/permission/matches-rule';
 import { ErrorCodes, KimiError } from '#/errors';
 import { z } from 'zod';
