+++ packages/agent-core/src/session/hooks/engine.ts
