+++ packages/agent-core/src/session/hooks/index.ts
