+++ packages/agent-core/src/session/hooks/types.ts
