+++ packages/agent-core/src/session/hooks/user-prompt.ts
